FinalProject
Application runs the game of boggle with three buttons(Start,Submit,Restart). Start displays the board and begins the countdown from two minutes. The player moves and score are displayed on the left of the board. As letters are selected from the board connections are drawn from the selected words displaying the currently selected word. If an invalid selection is made or the letter clicked cannot possibly form a word from the previous selected letters, then the currently selected word is reset. The Resources file must be in the same directory as the .jar file for the application to find the images. The user moves are displayed by a vbox with labels to display past moves. If too many moves are made the appended labels begin to go off screen but can be seen if the window is extended. There is not a way for the game to increase the chances a U is next to a Q. Some memory is not handled properly because when the restart button is clicked computer memory increases a lot but at some point stops increasing. This problem is still unresolved but most likely has to do with the way a new GameController is initialized on restart. Other than that there are no known bugs or issues with the game.